refactor : 패키지 구조와 클래스명을 명확하게 변경
