+++ detector/src/main/java/io/jeyong/detector/template/NPlusOneQueryCollector.java
@@ -1,13 +1,13 @@
 package io.jeyong.detector.template;
 
 import io.jeyong.detector.context.ExceptionContext;
-import io.jeyong.detector.test.NPlusOneQueryException;
+import io.jeyong.detector.exception.NPlusOneQueryException;
 
-public final class NPlusOneQueryExceptionCollector extends NPlusOneQueryTemplate {
+public final class NPlusOneQueryCollector extends NPlusOneQueryTemplate {
 
     private final ExceptionContext exceptionContext;
 
-    public NPlusOneQueryExceptionCollector(final int queryThreshold, final ExceptionContext exceptionContext) {
+    public NPlusOneQueryCollector(final int queryThreshold, final ExceptionContext exceptionContext) {
         super(queryThreshold);
         this.exceptionContext = exceptionContext;
     }
