+++ test/src/test/java/io/jeyong/test/mode/AnnotationLoggingModeTest.java
@@ -3,10 +3,10 @@ package io.jeyong.test.mode;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
+import io.jeyong.detector.annotation.NPlusOneTest;
 import io.jeyong.detector.config.NPlusOneDetectorProperties;
 import io.jeyong.detector.template.NPlusOneQueryLogger;
 import io.jeyong.detector.template.NPlusOneQueryTemplate;
-import io.jeyong.detector.test.NPlusOneTest;
 import io.jeyong.test.case2.service.ProductService;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
